requirements.txt

tensorflow==1.15
opencv-python
torch >= 1.7.1
torchvision
Flask==2.0.3
gunicorn==20.1.0
# Include other dependencies, e.g.:
numpy
Pillow
Werkzeug==2.0.3
